+++ x/core/01_interchain_security/module.go
@@ -1,16 +1,20 @@
 package interchain_security
 
 import (
+	"github.com/bcp-innovations/hyperlane-cosmos/x/core/01_interchain_security/client/cli"
 	"github.com/bcp-innovations/hyperlane-cosmos/x/core/01_interchain_security/types"
 	"github.com/cosmos/gogoproto/grpc"
 	"github.com/spf13/cobra"
 )
 
-import "github.com/bcp-innovations/hyperlane-cosmos/x/core/01_interchain_security/client/cli"
-
 // GetTxCmd returns the root command for the core ISMs
 func GetTxCmd() *cobra.Command {
 	return cli.GetTxCmd()
+}
+
+// GetQueryCmd returns the root command for the core ISMs
+func GetQueryCmd() *cobra.Command {
+	return cli.GetQueryCmd()
 }
 
 // RegisterMsgServer registers the core ism handler for transactions