+++ x/core/client/cli/tx.go
@@ -3,13 +3,12 @@ package cli
 import (
 	"fmt"
 
-	pdmodule "github.com/bcp-innovations/hyperlane-cosmos/x/core/02_post_dispatch"
-
+	"github.com/bcp-innovations/hyperlane-cosmos/x/core/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/spf13/cobra"
 
 	ism "github.com/bcp-innovations/hyperlane-cosmos/x/core/01_interchain_security"
-	"github.com/bcp-innovations/hyperlane-cosmos/x/core/types"
+	pdmodule "github.com/bcp-innovations/hyperlane-cosmos/x/core/02_post_dispatch"
 )
 
 var (